requirements.txt
openbb==4.*
python-dotenv==1.*
groq==0.*
